caraml-store-python/requirements-ci.txt
grpcio-tools==1.50.0
google-cloud-bigquery==3.0.0
protobuf==4.21.9
pandas==1.4.4